requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
sqlalchemy==2.0.23
aiohttp==3.8.6
reportlab==3.6.13
pydantic==2.4.2
langchain==0.0.267
python-dotenv==1.0.0
openai-whisper==20231117
torch==2.1.0
ffmpeg-python==0.2.0
transformers==4.51.3
einops==0.7.0
transformers_stream_generator==0.0.4
accelerate==0.27.2